% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajlFactorImputation-package.R
\docType{package}
\name{ajlFactorImputation-package}
\alias{ajlFactorImputation}
\alias{ajlFactorImputation-package}
\title{ajlFactorImputation: run exploratory factor analysis with multiple imputation for missing data values}
\description{
An easy to use interface to multiple imputation + factor anaylsis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AndrewLawrence/ajlFactorImputation}
}

}
\author{
\strong{Maintainer}: Andrew Lawrence \email{lawrenceajl@gmail.com}

}
\keyword{internal}
