% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loadings.R
\name{get_loadings}
\alias{get_loadings}
\title{Extract the loadings of the factor analysis,
usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.}
\usage{
get_loadings(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}
}
\description{
Extract the loadings of the factor analysis,
usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.
}
