% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mids.R
\name{get_mids}
\alias{get_mids}
\title{Extract the mice "mids" object, usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.}
\usage{
get_mids(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}
}
\description{
Extract the mice "mids" object, usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.
}
