% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covmat.R
\name{get_pooledcov}
\alias{get_pooledcov}
\title{Extract the \code{\link[mifa]{mifa}} pooled covariance matrix,
usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.}
\usage{
get_pooledcov(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}
}
\description{
Extract the \code{\link[mifa]{mifa}} pooled covariance matrix,
usually from a \code{\link{ajlfacimp-class}}
object, the output of \code{\link{factor_imputation}}.
}
