% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_imputation_options.R
\name{fami_options}
\alias{fami_options}
\title{fami_options}
\usage{
fami_options(
  fa_type = c("fa", "pca"),
  nfactors = "parallel",
  quickrun = FALSE,
  mice_var_checks = FALSE,
  n_imputations = NULL,
  ntree = NULL,
  maxit = NULL,
  defaultMethod = NULL
)
}
\arguments{
\item{fa_type}{Either Factor Analysis or PCA}

\item{nfactors}{Either, 1) the integer number of factors (or components)
to extract, -OR- 2) the string "parallel" indicating the parallel method
(\link[psych]{fa.parallel}) should be applied to estimate the number of
factors/components. The fa_type argument will determine whether factors
or components are used.}

\item{quickrun}{If TRUE, options will skimp on iterations to test basics}

\item{mice_var_checks}{If \code{FALSE} (default) will set arguments so mice
does not run particular variable checks (constants,
multicollinearity, high correlation) that are not relevant
when using random forests as the imputation model. For non-rf multiple
imputation methods, disabling checks may produce errors/poor
performance. The default imputation method in
\code{\link{ajlFactorImputation}} is random forest, so these checks are
disabled by default.}

\item{n_imputations}{Number of mice imputations (called \code{m} by mice).
If NULL, will be 10 (quickrun: 5)}

\item{ntree}{The number of trees to grow in the random forest imputation
method, default of 100 is used (quickrun: 10)}

\item{maxit}{A scalar giving the number of iterations. The default is 5.}

\item{defaultMethod}{How should mice impute values? This is a length 4
character vector (see \code{\link[mice]{mice}}) giving defaults for four
variable types. If unspecified (\code{NULL}) then random forest ("rf")
will be used for all types.}
}
\description{
Initialise options for \link{factor_imputation}. Detailed settings for
imputation and factor analysis are initialised, but the important overall
setting parameters are:
\itemize{
\item fa_type = do you want a PCA or a common factor model?
\item nfactors = how many factors to extract (or a method to calculate this)
\item quickrun = tries to set several parameters to run quickly
(if inaccurately) for testing purposes.
}
}
